Untangle headers
