+++ src/Functions/sparseGrams.cpp
@@ -1,7 +1,23 @@
-#include <Functions/sparseGrams.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionTokens.h>
+#include <Functions/sparseGramsImpl.h>
 
 namespace DB
 {
+
+namespace ErrorCodes
+{
+extern const int ILLEGAL_TYPE_OF_ARGUMENT;
+}
+
+/** Functions that finds all substrings win minimal length n
+  * such their border (n-1)-grams' hashes are more than hashes of every (n-1)-grams' in substring.
+  * As a hash function use zlib crc32, which is crc32-ieee with 0xffffffff as initial value
+  *
+  * sparseGrams(s)
+  */
+using FunctionSparseGrams = FunctionTokens<SparseGramsImpl<false>>;
+using FunctionSparseGramsUTF8 = FunctionTokens<SparseGramsImpl<true>>;
 
 REGISTER_FUNCTION(SparseGrams)
 {
@@ -54,52 +70,8 @@ Uses `CRC32` as a hash function.
     };
     FunctionDocumentation documentation_sparse_utf8 = {description_sparse_utf8, syntax_sparse_utf8, arguments_sparse, {}, returned_value_sparse_utf8, examples_sparse_utf8, introduced_in, category};
 
-    FunctionDocumentation::Description description_hashes = R"(
-Finds hashes of all substrings of a given string that have a length of at least `n`,
-where the hashes of the (n-1)-grams at the borders of the substring
-are strictly greater than those of any (n-1)-gram inside the substring.
-Uses `CRC32` as a hash function.
-)";
-    FunctionDocumentation::Syntax syntax_hashes = "sparseGramsHashes(s[, min_ngram_length, max_ngram_length])";
-    FunctionDocumentation::ReturnedValue returned_value_hashes = {"Returns an array of selected substrings CRC32 hashes.", {"Array(UInt32)"}};
-    FunctionDocumentation::Examples examples_hashes = {
-    {
-        "Usage example",
-        "SELECT sparseGramsHashes('alice', 3)",
-        R"(
-┌─sparseGramsHashes('alice', 3)──────────────────────┐
-│ [1481062250,2450405249,4012725991,1918774096]      │
-└────────────────────────────────────────────────────┘
-        )"
-    }
-    };
-    FunctionDocumentation documentation_hashes = {description_hashes, syntax_hashes, arguments_sparse, {}, returned_value_hashes, examples_hashes, introduced_in, category};
-
-    FunctionDocumentation::Description description_hashes_utf8 = R"(
-Finds hashes of all substrings of a given UTF-8 string that have a length of at least `n`, where the hashes of the (n-1)-grams at the borders of the substring are strictly greater than those of any (n-1)-gram inside the substring.
-Expects UTF-8 string, throws an exception in case of invalid UTF-8 sequence.
-Uses `CRC32` as a hash function.
-)";
-    FunctionDocumentation::Syntax syntax_hashes_utf8 = "sparseGramsHashesUTF8(s[, min_ngram_length, max_ngram_length])";
-    FunctionDocumentation::ReturnedValue returned_value_hashes_utf8 = {"Returns an array of selected UTF-8 substrings CRC32 hashes.", {"Array(UInt32)"}};
-    FunctionDocumentation::Examples examples_hashes_utf8 = {
-    {
-        "Usage example",
-        "SELECT sparseGramsHashesUTF8('алиса', 3)",
-        R"(
-┌─sparseGramsHashesUTF8('алиса', 3)─┐
-│ [4178533925,3855635300,561830861] │
-└───────────────────────────────────┘
-        )"
-    }
-    };
-    FunctionDocumentation documentation_hashes_utf8 = {description_hashes_utf8, syntax_hashes_utf8, arguments_sparse, {}, returned_value_hashes_utf8, examples_hashes_utf8, introduced_in, category};
-
     factory.registerFunction<FunctionSparseGrams>(documentation_sparse);
     factory.registerFunction<FunctionSparseGramsUTF8>(documentation_sparse_utf8);
-
-    factory.registerFunction<SparseGramsHashes<false>>(documentation_hashes);
-    factory.registerFunction<SparseGramsHashes<true>>(documentation_hashes_utf8);
 }
 
 }
